# Faster MLA prompt processing