@ikawrakow 2025-02-13
The PR also adds a compile time option to disable the transposed KV cache when using MLA (simple look for `MLA_USE_TRANSPOSED_CACHE` and set it to 0). This cuts KV cache size in nearly half at the expense of a lower TG performance with long contexts. PP performance stays about the same. Here is a comparison between MLA with and without transposed cache

 | model                |          test |  t/s (with c^T)  |  t/s (without c^T)|
| -------------------- | ------------: | ---------------: | ----------------: |
| deepseek2 16B IQ4_XS |    tg64@pp128 |     33.58 ± 0.06 |      33.05 ± 0.05 |
| deepseek2 16B IQ4_XS |    tg64@pp256 |     32.67 ± 0.00 |      31.54 ± 0.07 |
| deepseek2 16B IQ4_XS |    tg64@pp512 |     32.38 ± 0.08 |      30.26 ± 0.33 |
| deepseek2 16B IQ4_XS |   tg64@pp1024 |     31.50 ± 0.02 |      28.50 ± 0.01 |
| deepseek2 16B IQ4_XS |   tg64@pp2048 |     30.01 ± 0.01 |      24.75 ± 0.01 |
| deepseek2 16B IQ4_XS |   tg64@pp4096 |     27.08 ± 0.03 |      20.67 ± 0.09 |
| deepseek2 16B IQ4_XS |   tg64@pp8192 |     22.82 ± 0.00 |      14.89 ± 0.01 |